refactoring code by moving common methods to drutils file
